server/requirements.txt
# DataCleaner OpenEnv — runtime dependencies
# This file is the single source of truth used by the Dockerfile.
# Keep in sync with pyproject.toml [project.dependencies].
openenv-core[core]>=0.2.2
openai>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
rapidfuzz>=3.0.0
uvicorn[standard]>=0.20.0
fastapi>=0.100.0
